requirements.txt
python-telegram-bot==13.6
requests==2.25.1
beautifulsoup4==4.9.3
pyTelegramBotAPI==3.8.1
psycopg2-binary==2.9.1
python-dotenv==0.18.0